Moved config reading functions to config file. Documented functions related to TLE reading.
